+++ api/src/main/java/ru/geekbraines/spring/winter/market/api/ProductDto.java
@@ -1,7 +1,6 @@
 package ru.geekbraines.spring.winter.market.api;
 
-
-
+import java.math.BigDecimal;
 
 //@NoArgsConstructor // джесоновские объекты должны быть всегда с дефолтным конструктором (джексон, когда получает/отдает объект
 // в джейсоне, использует дефолтный конструктор для его формирования)
@@ -13,7 +12,7 @@ public class ProductDto {
 
     private Long id;   
     private String title;
-    private int price;
+    private BigDecimal price;
     private String categoryTitle;
 
     
@@ -29,10 +28,10 @@ public class ProductDto {
     public void setTitle(String title) {
         this.title = title;
     }
-    public int getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
-    public void setPrice(int price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
     public String getCategoryTitle() {
@@ -41,7 +40,7 @@ public class ProductDto {
     public void setCategoryTitle(String categoryTitle) {
         this.categoryTitle = categoryTitle;
     }
-    public ProductDto(Long id, String title, int price, String categoryTitle) {
+    public ProductDto(Long id, String title, BigDecimal price, String categoryTitle) {
         this.id = id;
         this.title = title;
         this.price = price;
